+++ nengo/tests/test_neurons.py
@@ -1,5 +1,8 @@
+from collections import defaultdict
 from inspect import getfullargspec
+import itertools
 import logging
+import time
 
 import numpy as np
 import pytest
@@ -15,9 +18,12 @@ from nengo.neurons import (
     LIFRate,
     NeuronType,
     NeuronTypeParam,
+    PoissonSpiking,
     RectifiedLinear,
+    RegularSpiking,
     Sigmoid,
     SpikingRectifiedLinear,
+    StochasticSpiking,
     Tanh,
 )
 from nengo.processes import WhiteSignal
@@ -25,6 +31,16 @@ from nengo.solvers import LstsqL2nz
 from nengo.utils.ensemble import tuning_curves
 from nengo.utils.matplotlib import implot, rasterplot
 from nengo.utils.numpy import rms
+
+
+# --- define a composite neuron type, used in `nl` (see `pytest_nengo.py`)
+class SpikingTanh(RegularSpiking):
+    def __init__(self, tau_ref=0.0025, amplitude=1.0):
+        super().__init__(base_type=Tanh(tau_ref=tau_ref), amplitude=amplitude)
+
+    @property
+    def tau_ref(self):
+        return self.base_type.tau_ref
 
 
 def test_lif_builtin(rng, allclose):
@@ -336,6 +352,57 @@ def test_sigmoid_invalid(Simulator, max_rate, intercept):
     with pytest.raises(BuildError):
         with Simulator(m):
             pass
+
+
+@pytest.mark.parametrize("base_rate", [LIFRate(), RectifiedLinear(), Tanh()])
+def test_spiking_types(base_rate, Simulator, seed, plt, allclose):
+    spiking_types = {
+        RegularSpiking: dict(atol=0.05, rmse_target=0.011),
+        PoissonSpiking: dict(atol=0.13, rmse_target=0.024),
+        StochasticSpiking: dict(atol=0.10, rmse_target=0.019),
+    }
+
+    n_neurons = 1000
+
+    with nengo.Network(seed=seed) as net:
+        u = nengo.Node(lambda t: np.sin(2 * np.pi * t))
+        a = nengo.Ensemble(n_neurons, 1)
+        nengo.Connection(u, a)
+        u_p = nengo.Probe(u, synapse=0.005)
+        a_p = nengo.Probe(a, synapse=0.005)
+
+    neuron_types = {
+        spiking_type: spiking_type(base_rate) for spiking_type in spiking_types
+    }
+    neuron_types[None] = base_rate
+
+    delay = 5
+    results = defaultdict(dict)
+    for neuron_type in neuron_types.values():
+        a.neuron_type = neuron_type
+
+        with nengo.Simulator(net, seed=seed + 1) as sim:
+            timer = time.time()
+            sim.run(1.0)
+            timer = time.time() - timer
+
+        results[neuron_type]["u"] = sim.data[u_p]
+        results[neuron_type]["x"] = sim.data[a_p]
+        plt.plot(sim.trange(), sim.data[a_p], label="%s: t=%.3f" % (neuron_type, timer))
+
+    plt.plot(sim.trange()[delay:], sim.data[u_p][:-delay], "k--")
+    plt.legend(loc=3)
+
+    for spiking_type, tols in spiking_types.items():
+        neuron_type = neuron_types[spiking_type]
+        res = results[neuron_type]
+        x = res["x"][delay:]
+        u = res["u"][:-delay]
+        assert allclose(x, u, atol=tols["atol"]), spiking_type
+
+        # check that spike noise is the target amount above the base spiking model noise
+        rmse = rms(x - u)
+        assert allclose(rmse, tols["rmse_target"], atol=0.003, rtol=0.25), spiking_type
 
 
 def test_dt_dependence(Simulator, nl_nodirect, plt, seed, allclose):
@@ -702,3 +769,18 @@ def test_argreprs():
     check_repr(
         Izhikevich(tau_recovery=0.1, coupling=0.3, reset_voltage=-1, reset_recovery=5)
     )
+
+    comp_types = {
+        RegularSpiking: [dict(), dict(amplitude=0.4)],
+        PoissonSpiking: [dict(), dict(amplitude=0.2)],
+        StochasticSpiking: [dict(), dict(amplitude=0.3)],
+    }
+    base_types = {
+        RectifiedLinear: [dict(), dict(amplitude=0.7)],
+        Sigmoid: [dict(), dict(tau_ref=0.03)],
+        Tanh: [dict(), dict(tau_ref=0.03)],
+    }
+    for comp_type, comp_params in comp_types.items():
+        for base_type, base_params in base_types.items():
+            for comp_kwargs, base_kwargs in itertools.product(comp_params, base_params):
+                check_repr(comp_type(base_type(**base_kwargs), **comp_kwargs))